+++ src/cw/op/op_lld.c
@@ -7,7 +7,7 @@ void			op_lld(t_cw *cw, size_t i_car)
 	if (cw->cod.arg.v1 == IND_CODE)
 	{
 		cw->pos = cw->car[i_car]->pc + OPC_SIZE + IND_SIZE;
-		cw->arg[0] = code_to_byte(cw->map, cw->pos, IND_SIZE) + cw->car[i_car]->pc;
+		cw->arg[0] = (int16_t)code_to_byte(cw->map, cw->pos, IND_SIZE) + cw->car[i_car]->pc;
 	}
 	cw->car[i_car]->reg[IN(cw->arg[1])] = code_to_byte(cw->map, cw->arg[0], REG_SIZE);
 	cw->car[i_car]->carry = (cw->car[i_car]->reg[IN(cw->arg[1])]) ? 0 : 1;
